// apps/backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserType = {
  CUSTOMER: 'CUSTOMER'
} as const

export type UserType = (typeof UserType)[keyof typeof UserType]


export const ItemType = {
  MOVIE: 'MOVIE',
  SERIES: 'SERIES'
} as const

export type ItemType = (typeof ItemType)[keyof typeof ItemType]


export const ActionType = {
  VIEW: 'VIEW',
  LIKE: 'LIKE',
  DISLIKE: 'DISLIKE',
  RATING: 'RATING',
  FAVORITE: 'FAVORITE',
  SEARCH: 'SEARCH'
} as const

export type ActionType = (typeof ActionType)[keyof typeof ActionType]